requirements.txt
numpy>=1.26.0
anthropic>=0.42.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
python-multipart>=0.0.18
pydantic>=2.0.0
aiosqlite>=0.20.0   # async SQLite driver (required for SQLAlchemy + aiosqlite backend)
aiofiles>=23.0.0    # async file writes for the uploads route
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
passlib[bcrypt]
